# Reduce objects allocation in the java heap