## Context

- Implement `BooleanArrayPool`. It supposed to reduce object allocation rate;

Some benchmarks from my Apple MacBook Air M1:
```
# JMH version: 1.36
# VM version: JDK 17.0.1, OpenJDK 64-Bit Server VM, 17.0.1+12
# VM invoker: .../17.0.1-tem/bin/java

# NEW
Benchmark                                                      Mode  Cnt         Score   Error   Units
JaroWinklerSimilarityBenchmark.benchmark                      thrpt   15         1.437 ± 0.010   ops/s
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate       thrpt   15        54.833 ± 0.365  MB/sec
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate.norm  thrpt   15  40004687.680 ± 4.442    B/op
JaroWinklerSimilarityBenchmark.benchmark:·gc.count            thrpt   15        15.000          counts
JaroWinklerSimilarityBenchmark.benchmark:·gc.time             thrpt   15        10.000              ms

# OLD
Benchmark                                                      Mode  Cnt          Score   Error   Units
JaroWinklerSimilarityBenchmark.benchmark                      thrpt   15          1.389 ± 0.023   ops/s
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate       thrpt   15        197.755 ± 3.216  MB/sec
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate.norm  thrpt   15  149300361.054 ± 5.295    B/op
JaroWinklerSimilarityBenchmark.benchmark:·gc.count            thrpt   15         44.000          counts
JaroWinklerSimilarityBenchmark.benchmark:·gc.time             thrpt   15         31.000              ms

```

## Checklist
- [x] Change meets or does not compromise the [Baseline Security Requirements](https://transferwise.atlassian.net/wiki/spaces/EKB/pages/434929973/Baseline+Security+Requirements)